Aggregate test reports of suite replay results
Renamed `SuiteReplayResultProvider` in `SuiteAggregator`, because it returns the aggregated test report of all suite replay results and not a single suite anymore. Test report is not static anymore, because `SuiteAggregator` is already a singleton
wrapper ctor for `TestReport#fromAPI()` with `GoldenMasterSource.API` outsourced hardcoded name for test report file into constants in `Properties`.