Add a whitelist for the important metadata differences
as potentially only a few metadata is really relevant for the user, i.e. that have an effect on the resulting golden master. This would then mean, however, that we would include all web specific keys here, but I would argue that this is nothing new in this repo.

Otherwise we would have to integrate a flag or similar to disable the printing of the metadata, since we have some tests that use approvals to test the output produced by a recheck test. This would go as far as to make this flag