Do not try to evaluate project root five times
instead evaluate once and reuse locally.. We do not want to cache the project as the frontend or the api should be responsible for this. This is, so that the system in question can define the project root manually (e,g. opening a report with no project and setting the project to the corresponding folder).