Specify new method to retrieve the test source root from the ProjectLayout
The method returns an Optional, since we strictly do not need it per se, we just cannot perform test healing. Secondly, this is for compatibility, since this is public api and we do not want to break the existing implementations but do want to return something more sensible than null.