Add classes for highlighting
This package is added to enable highlighting in recheck.cli.
The printers for test reports are hard to read despite formatting
improvements. This commit adds the capability to highlight certain
keywords (which may be expanded at any time, as needed).
Because of the design of the printers, it was necessary to create a
global highlighter.
The interface highlighter may be implemented to create a
custom highlighter. By default, the DefaultHighlighter is used, which
just returns the original string without any modifications.