Move reading report data into metadata to top
to have better control over exceptions and log properly if the file could not be read.
This honors if the report does not have any differences and warns or rethrows accordingly.
Secondly, it does not attempt to re-read the file, because this should be a rare case and usually indicates some other severe problem which cannot be fixed with a simple retry.