Create a separate warnings lookup
which is necessary, since it might happen that there are two similar differences, one with warnings and the other without. When accepting, it might be possible that the difference without the warning is accepted, and per 1-click-maintenance the other is accepted as well. However, only the passed AttributeDifference is added to the changes, thus the warning is lost.

Thus the AttributeDifferences is reconstructed by injecting any found warning prior to adding or removing them from the accepted changes.

Strictly speaking, this is validates the accepted changes, because the difference accepted cannot be found in the corresponding element difference (as a AttributeDifference with and without warning are different). However, we never really check this and thus it should have no effect on the apply mechanism.