Lazily detect the project layout
but ensure it is set first thing within RecheckOptionsBuilder#build.
Firstly, this is done to provide a meaningful stacktrace when the project layout could not be found
and secondly to allow other lazily initialized methods to access it.

However, due to how project discovery works, the project found may not be the execution directory. To work around this, the test must create more indicator files (e.g. src).